Pattern data structure redesign.
pattern.py no longer does the conversion into the weird data format of
the bemis100 itself (expressing data by counting the number of ones in a
byte). Instead, the encoding of each character should be handled by
bemis100.py. This makes it easier to have general patterns that play on
the GE lights as well.