Reconfigured to always specify the number of lights, not boards
We have always talked about the number of "boards", each of which
contains two independent RGB pixels. This gets confusing when we deal
with other hardware, though, so I'm switching to num_lights being the
common variable for patterns and controllers. This is ~166 for the
bemis100 and 50 for the GE lights