Removed subclassing of LEDController.
Rather than having a Bemis100Controller or a GEController, we now just create a single LEDController instance and add Bemis100Writers and GEWriters to it. This makes it easy to logically separate the controller from the hardware, and will be useful for my next goal of allowing the attaching and detaching of hardware from the web interface.